+++ components/src/core/drawer/drawer-layout/drawer-layout.component.ts
@@ -9,6 +9,7 @@ import {
     Output,
     ViewChild,
     ViewEncapsulation,
+    SimpleChanges,
 } from '@angular/core';
 import { DrawerService } from '../service/drawer.service';
 import { StateListener } from '../state-listener.component';
@@ -82,6 +83,10 @@ export class DrawerLayoutComponent extends StateListener implements AfterViewIni
     @ViewChild('remElement') remElement: ElementRef;
 
     isRtl = false;
+
+    /** This is true whenever the drawer is in a collapsed state & its variant has been transitioned to temporary. */
+    hasCollapsedTransitionToTemporary = false;
+
     remSizePx: number;
     dirChangeSubscription = Subscription.EMPTY;
 
@@ -103,8 +108,24 @@ export class DrawerLayoutComponent extends StateListener implements AfterViewIni
         this.isRtl = this._dir.value === 'rtl';
     }
 
-    ngOnChanges(): void {
+    ngOnChanges(simpleChanges: SimpleChanges): void {
         this.drawerService.setDrawerVariant(this.variant);
+
+        // Whenever a drawer has transitioned from a closed persistent drawer to a temporary variant,
+        // this edge case prevents the drawer from being redrawn at an opened size before it dismissed.
+        if (simpleChanges && simpleChanges.variant) {
+            const variant = simpleChanges.variant;
+            if (
+                variant.currentValue === 'temporary' &&
+                (variant.previousValue === 'rail' || (variant.previousValue === 'persistent' && !this.isOpen()))
+            ) {
+                this.hasCollapsedTransitionToTemporary = true;
+                setTimeout(() => {
+                    this.hasCollapsedTransitionToTemporary = false;
+                }, 500);
+            }
+        }
+
         this.changeDetector.detectChanges();
     }
 
@@ -157,6 +178,7 @@ export class DrawerLayoutComponent extends StateListener implements AfterViewIni
     isCollapsed(): boolean {
         if (this.variant === 'rail') return true; // Rail is always collapsed.
         if (this.variant === 'persistent') return !this.isOpen();
+        if (this.hasCollapsedTransitionToTemporary) return true;
         return false;
     }
 }